--- a/scripts/utils/trading_strategy_helpers.js
+++ b/scripts/utils/trading_strategy_helpers.js
@@ -605,10 +605,8 @@ const buildTransferFundsToMaster = async function(masterAddress, withdrawals, li
  * @return {Transaction} Multisend transaction that has to be sent from the master address to transfer back the funds stored in the exchange
  */
 const buildWithdrawAndTransferFundsToMaster = async function(masterAddress, withdrawals, web3 = web3, artifacts = artifacts) {
-  const [withdrawalTransaction, transferFundsToMasterTransaction] = await Promise.all([
-    buildWithdraw(masterAddress, withdrawals, web3, artifacts),
-    buildTransferFundsToMaster(masterAddress, withdrawals, false, web3, artifacts)
-  ])
+  const withdrawalTransaction = await buildWithdraw(masterAddress, withdrawals, web3, artifacts)
+  const transferFundsToMasterTransaction = await buildTransferFundsToMaster(masterAddress, withdrawals, false, web3, artifacts)
   return buildBundledTransaction([withdrawalTransaction, transferFundsToMasterTransaction], web3, artifacts)
 }