--- a/scripts/utils/trading_strategy_helpers.js
+++ b/scripts/utils/trading_strategy_helpers.js
@@ -88,7 +88,7 @@ const maxUINT = new BN(2).pow(new BN(256)).sub(new BN(1))
 /**
  * Returns an instance of the exchange contract
  */
-const getExchange = async function(web3) {
+const getExchange = function(web3) {
   BatchExchange.setNetwork(web3.network_id)
   BatchExchange.setProvider(web3.currentProvider)
   return BatchExchange.deployed()