… test/dfusionMultiSafes.js