Drastically reduce the running time of the withdrawal scripts by heavily relying on Javascript's asynchronous features.

This PR is a test bench for using the same techniques everywhere else in this repo, so changes and recommendations are very welcome before we start working on the other scripts.

Notable breaking change is the function `fetchTokenInfo`, which is now split in `fetchTokenInfoAtAddresses` and `fetchTokenInfoFromExchange`, both of which return a promise.

Running times of withdraw.js
1. with two safes
  - before: 28±1s
  - after: 12±1s

2. with 20 safes
  - before: 3min 6s (no ±, I ran it only once)
  - after: 14.7±0.1s


Related issue: #57 
Thanks @nlordell for the discussion.