@fedgiac 2020-03-13
Thanks for the comments! I also noticed that these changes introduce a bug in the logging: the function console.log.apply is apparently not thread safe, and the result is that instead of returning some output it just prints an empty line.
Since logging is a bit messing currently, I'd say we could merge anyway and then take care of the logging in another PR.