@fedgiac 2020-03-13
I checked a previous commit: logging was already broken back then in the same way as it is with this PR, this PR (as well as the previous one changing `fetchTokenInfo`) didn't introduce this problem.